Bluetooth: MAP removed opened cursor.
- Replaced opened cursor with collections.
- Restructured classes to use table for MAS instances.
- Restructured MAP message handle offset for multiple MAS instances.
- MAS instance for Email uses the primary email account at the time
   of connection and continue to use it for that connection
   unless it is removed.

CRs-fixed: 311515

Conflicts:

    src/com/android/bluetooth/map/BluetoothMasService.java
--
Bluetooth: MAP work around for BMW carkit

 - Ignore improper filtering request for GetMessageListing

CRs-fixed: 315435
--
Bluetooth: MAP prepare for 2nd email account support

- Fixed hardcoded mas instance id/email account id

--
Bluetooth: MAP Multiple MAS instances.

- Fixed MAS listening thread despawn after wrongful connect attempt

CRs-fixed: 306061
--

+++ src/com/android/bluetooth/map/MapUtils/MsgListingConsts.java
@@ -44,7 +44,7 @@ public class MsgListingConsts {
     }
     public MsgInfo msgInfo = new MsgInfo();
 
-    public int msg_handle = 0;
+    public long msg_handle = 0;
     public String subject = null;
     public boolean  sendSubject = false;
     public String datetime = null;
@@ -64,11 +64,11 @@ public class MsgListingConsts {
     public String msg_protected = null;
     public String replyto_addressing = null;
 
-    public int getMsg_handle() {
+    public long getMsg_handle() {
         return msg_handle;
     }
 
-    public void setMsg_handle(int msg_handle) {
+    public void setMsg_handle(long msg_handle) {
         this.msg_handle = msg_handle;
     }
